+++ exchangis-datasource/exchangis-datasource-loader/src/main/java/com/webank/wedatasphere/exchangis/datasource/loader/loader/LocalExchangisDataSourceLoader.java
@@ -7,6 +7,7 @@ import com.webank.wedatasphere.exchangis.datasource.core.loader.ExchangisDataSou
 import com.webank.wedatasphere.exchangis.datasource.loader.clazzloader.ExchangisDataSourceClassLoader;
 import com.webank.wedatasphere.exchangis.datasource.loader.utils.ExceptionHelper;
 import com.webank.wedatasphere.exchangis.datasource.loader.utils.ExtDsUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.apache.linkis.common.exception.ErrorException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -38,7 +39,9 @@ public class LocalExchangisDataSourceLoader implements ExchangisDataSourceLoader
         // 初始化磁盘扫描加载
         ClassLoader currentClassLoader = Thread.currentThread().getContextClassLoader();
         String loadClassPath =  Objects.requireNonNull(currentClassLoader.getResource(EXCHANGIS_DIR_NAME)).getPath();
-//        String libPathUrl = loadClassPath + ".." + File.separator + ".." + File.separator + EXCHANGIS_DIR_NAME;
+        if (StringUtils.endsWith(loadClassPath, File.separator)) {
+            loadClassPath = loadClassPath + File.separator;
+        }
         String libPathUrl = loadClassPath + ".." + File.separator + EXCHANGIS_DIR_NAME;
         LOGGER.info("libPath url is {}",  libPathUrl);
         List<URL> jars = ExtDsUtils.getJarsUrlsOfPath(libPathUrl);
