+++ exchangis-engines/engineconn-plugins/datax/src/main/scala/org/apache/linkis/engineconnplugin/datax/DataxEngineConnPlugin.scala
@@ -26,7 +26,9 @@ import org.apache.linkis.manager.engineplugin.common.launch.EngineConnLaunchBuil
 import org.apache.linkis.manager.engineplugin.common.resource.EngineResourceFactory
 import org.apache.linkis.manager.label.entity.Label
 
-import java.util
+import java.util.Map
+import java.util.List
+import java.util.ArrayList
 
 class DataxEngineConnPlugin extends EngineConnPlugin {
 
@@ -36,7 +38,7 @@ class DataxEngineConnPlugin extends EngineConnPlugin {
   private var engineConnFactory: EngineConnFactory = _
   private val engineConnFactoryLocker = new Array[Byte](0)
 
-  override def init(params: util.Map[String, Any]): Unit = {}
+  override def init(params: Map[String, AnyRef]): Unit = {}
 
   override def getEngineResourceFactory: EngineResourceFactory = {
     if (null == engineResourceFactory) engineResourceFactoryLocker.synchronized {
@@ -60,7 +62,7 @@ class DataxEngineConnPlugin extends EngineConnPlugin {
     engineConnFactory
   }
 
-  override def getDefaultLabels: util.List[Label[_]] = {
-    new util.ArrayList[Label[_]]()
+  override def getDefaultLabels: List[Label[_]] = {
+    new ArrayList[Label[_]]()
   }
 }
