+++ exchangis-engines/engineconn-plugins/sqoop/pom.xml
@@ -32,7 +32,7 @@
         <sqoop.version>1.4.6</sqoop.version>
         <hive.version>3.1.2</hive.version>
         <hadoop.version>2.7.2</hadoop.version>
-        <linkis.version>1.1.6-webank</linkis.version>
+        <linkis.version>1.4.0-SNAPSHOT</linkis.version>
     </properties>
 
     <dependencies>
@@ -187,28 +187,24 @@
             <artifactId>linkis-engineconn-plugin-core</artifactId>
             <version>${linkis.version}</version>
         </dependency>
-
         <dependency>
             <groupId>org.apache.linkis</groupId>
             <artifactId>linkis-rpc</artifactId>
             <version>${linkis.version}</version>
             <scope>provided</scope>
         </dependency>
-
         <dependency>
             <groupId>org.apache.linkis</groupId>
             <artifactId>linkis-storage</artifactId>
             <version>${linkis.version}</version>
             <scope>provided</scope>
         </dependency>
-
         <dependency>
             <groupId>org.apache.linkis</groupId>
             <artifactId>linkis-common</artifactId>
             <version>${linkis.version}</version>
             <scope>provided</scope>
         </dependency>
-
     </dependencies>
 
     <build>
@@ -217,7 +213,6 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-deploy-plugin</artifactId>
             </plugin>
-
             <plugin>
                 <groupId>net.alchim31.maven</groupId>
                 <artifactId>scala-maven-plugin</artifactId>
