+++ exchangis-engines/engineconn-plugins/sqoop/src/main/scala/org/apache/linkis/engineconnplugin/sqoop/SqoopEngineConnPlugin.scala
@@ -25,16 +25,17 @@ import org.apache.linkis.manager.label.entity.Label
 import org.apache.linkis.engineconnplugin.sqoop.factory.SqoopEngineConnFactory
 import org.apache.linkis.engineconnplugin.sqoop.launch.SqoopEngineConnLaunchBuilder
 
+import java.util.Map
+import java.util.List
 
 class SqoopEngineConnPlugin extends EngineConnPlugin{
   private val EP_CONTEXT_CONSTRUCTOR_LOCK = new Object()
   private var engineResourceFactory: EngineResourceFactory = _
   private var engineConnLaunchBuilder: EngineConnLaunchBuilder = _
   private var engineConnFactory: EngineConnFactory = _
-  override def init(params: java.util.Map[String, Any]): Unit = {}
+  override def init(params: Map[String, AnyRef]): Unit = {}
 
   override def getEngineResourceFactory: EngineResourceFactory = {
-
     EP_CONTEXT_CONSTRUCTOR_LOCK.synchronized{
       if(null == engineResourceFactory){
         engineResourceFactory = new GenericEngineResourceFactory
@@ -62,5 +63,6 @@ class SqoopEngineConnPlugin extends EngineConnPlugin{
     }
   }
 
-  override def getDefaultLabels: java.util.List[Label[_]] = new java.util.ArrayList[Label[_]]
+  override def getDefaultLabels: List[Label[_]] = new java.util.ArrayList[Label[_]]
+
 }
