+++ pom.xml
@@ -38,11 +38,12 @@
 
 	<properties>
 		<exchangis.version>1.1.2</exchangis.version>
-		<dss.version>1.1.0</dss.version>
-		<linkis.version>1.1.6-webank</linkis.version>
-		<linkis.datasource.version>1.1.1</linkis.datasource.version>
+		<dss.version>1.1.2</dss.version>
+		<linkis.version>1.4.0-SNAPSHOT</linkis.version>
+		<linkis.datasource.version>1.4.0-SNAPSHOT</linkis.datasource.version>
 		<streamis.version>0.1.0-SNAPSHOT</streamis.version>
-		<scala.version>2.11.8</scala.version>
+		<scala.version>2.12.17</scala.version>
+		<scala-maven-plugin.version>4.7.1</scala-maven-plugin.version>
 		<jdk.compile.version>1.8</jdk.compile.version>
 		<maven.version>3.3.3</maven.version>
 		<gson.version>2.8.5</gson.version>
@@ -66,7 +67,6 @@
 		<xstream.version>1.4.19</xstream.version>
 		<jobmanager.version>0.1.0-SNAPSHOT</jobmanager.version>
 	</properties>
-
 
 	<modules>
 		<module>exchangis-dao</module>
@@ -109,10 +109,17 @@
 				<groupId>org.apache.linkis</groupId>
 				<artifactId>linkis-mybatis</artifactId>
 				<version>${linkis.version}</version>
+				<exclusions>
+					<exclusion>
+						<groupId>org.springframework</groupId>
+						<artifactId>spring-orm</artifactId>
+					</exclusion>
+				</exclusions>
 			</dependency>
 			<dependency>
 				<groupId>org.apache.linkis</groupId>
 				<artifactId>linkis-module</artifactId>
+				<version>${linkis.version}</version>
 				<exclusions>
 					<exclusion>
 						<groupId>org.springframework.boot</groupId>
@@ -123,20 +130,22 @@
 						<groupId>org.hibernate.validator</groupId>
 					</exclusion>
 				</exclusions>
-				<version>${linkis.version}</version>
 			</dependency>
 			<dependency>
 				<groupId>org.apache.linkis</groupId>
 				<artifactId>linkis-common</artifactId>
 				<version>${linkis.version}</version>
 			</dependency>
-
 			<dependency>
 				<groupId>org.apache.linkis</groupId>
 				<artifactId>linkis-protocol</artifactId>
 				<version>${linkis.version}</version>
 			</dependency>
-
+			<dependency>
+				<groupId>org.apache.linkis</groupId>
+				<artifactId>linkis-datasource-client</artifactId>
+				<version>${linkis.datasource.version}</version>
+			</dependency>
 			<dependency>
 				<groupId>com.google.code.gson</groupId>
 				<artifactId>gson</artifactId>
@@ -162,11 +171,10 @@
 				<groupId>com.thoughtworks.xstream</groupId>
 				<version>${xstream.version}</version>
 			</dependency>
-
 			<dependency>
-				<groupId>org.apache.linkis</groupId>
-				<artifactId>linkis-datasource-client</artifactId>
-				<version>${linkis.datasource.version}</version>
+				<groupId>org.springframework</groupId>
+				<artifactId>spring-orm</artifactId>
+				<version>5.2.15.RELEASE</version>
 			</dependency>
 		</dependencies>
 	</dependencyManagement>
@@ -225,7 +233,7 @@
 				<plugin>
 					<groupId>net.alchim31.maven</groupId>
 					<artifactId>scala-maven-plugin</artifactId>
-					<version>3.2.2</version>
+					<version>${scala-maven-plugin.version}</version>
 					<executions>
 						<execution>
 							<id>eclipse-add-source</id>
@@ -258,7 +266,6 @@
 					<configuration>
 						<scalaVersion>${scala.version}</scalaVersion>
 						<recompileMode>incremental</recompileMode>
-						<useZincServer>true</useZincServer>
 					</configuration>
 				</plugin>
 				<plugin>
